curve

+++ GrainPF_ex/ML_Grain/curve/main.cpp
@@ -546,7 +546,7 @@ int main(int argc, char** argv)
 
 
     for (int i=0; i<params.num_theta; i++){
-        grain_angle[i] = grain_grid[i]/radius_n;
+        grain_angle[i] = (float) grain_grid[i]/radius_n;
         printf("grain %d, PF %d, angle %f, grid %d, frac %f, arc %f\n",i,aseq[i],mac.theta_arr[i+1]/M_PI*180,grain_grid[i],frac_ini[i], grain_angle[i] );
     }
     float Dx = mac.X_mac[mac.Nx-1] - mac.X_mac[mac.Nx-2];
